@@ -21,7 +21,7 @@
 		"embla-carousel-autoplay": "^8.3.0",
 		"esbuild": "^0.24.0",
 		"flowbite": "1.8.1",
-		"svelte": "^4.2.19",
+		"svelte": "^5.14.4",
 		"tailwindcss": "^3.4.17",
 		"typescript": "^5.6.3"
 	},
