@@ -6,8 +6,14 @@
 
 	type $$Props = HTMLAttributes<HTMLDivElement>;
 
-	let className: string | undefined | null = undefined;
-	export { className as class };
+	interface Props {
+		class?: string | undefined | null;
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, children, ...rest }: Props = $props();
+	
 
 	const { orientation, options, plugins, onInit } = getEmblaContext("<Carousel.Content/>");
 </script>
@@ -23,13 +29,13 @@
 		},
 		plugins: $plugins,
 	}}
-	on:emblaInit={onInit}
+	onemblaInit={onInit}
 >
 	<div
 		class={cn("flex", $orientation === "horizontal" ? "-ml-4" : "-mt-4 flex-col", className)}
 		data-embla-container=""
-		{...$$restProps}
+		{...rest}
 	>
-		<slot />
+		{@render children?.()}
 	</div>
 </div>
