@@ -4,8 +4,14 @@
 	import { cn } from "@/utils.js";
 
 	type $$Props = HTMLAttributes<HTMLDivElement>;
-	let className: string | undefined | null = undefined;
-	export { className as class };
+	interface Props {
+		class?: string | undefined | null;
+		children?: import('svelte').Snippet;
+		[key: string]: any
+	}
+
+	let { class: className = undefined, children, ...rest }: Props = $props();
+	
 
 	const { orientation } = getEmblaContext("<Carousel.Item/>");
 </script>
@@ -19,7 +25,7 @@
 		className
 	)}
 	data-embla-slide=""
-	{...$$restProps}
+	{...rest}
 >
-	<slot />
+	{@render children?.()}
 </div>
