@@ -3,18 +3,23 @@
 	import type { VariantProps } from "tailwind-variants";
 	import { getEmblaContext } from "./context.js";
 	import { cn } from "@/utils.js";
-	import {
-		Button,
-		type Props,
-		type buttonVariants,
-	} from "@/components/ui/button/index.js";
+	import { Button, type buttonVariants } from "@/components/ui/button/index.js";
 
 	type $$Props = Props;
 
-	let className: $$Props["class"] = undefined;
-	export { className as class };
-	export let variant: VariantProps<typeof buttonVariants>["variant"] = "outline";
-	export let size: VariantProps<typeof buttonVariants>["size"] = "icon";
+	interface Props {
+		class?: $$Props["class"];
+		variant?: VariantProps<typeof buttonVariants>["variant"];
+		size?: VariantProps<typeof buttonVariants>["size"];
+		[key: string]: any;
+	}
+
+	let {
+		class: className = undefined,
+		variant = "outline",
+		size = "icon",
+		...rest
+	}: Props = $props();
 	const { orientation, canScrollNext, scrollNext, handleKeyDown } =
 		getEmblaContext("<Carousel.Next/>");
 </script>
@@ -27,12 +32,12 @@
 		$orientation === "horizontal"
 			? "-right-12 top-1/2 -translate-y-1/2"
 			: "-bottom-12 left-1/2 -translate-x-1/2 rotate-90",
-		className
+		className,
 	)}
 	disabled={!$canScrollNext}
 	on:click={scrollNext}
 	on:keydown={handleKeyDown}
-	{...$$restProps}
+	{...rest}
 >
 	<ArrowRight class="h-4 w-4" />
 	<span class="sr-only">Next slide</span>
